+++ Civi/ActionProvider/Action/Contact/GetEmailAddress.php
@@ -10,17 +10,18 @@ use \Civi\ActionProvider\Parameter\Specification;
 use \Civi\ActionProvider\Parameter\OptionGroupSpecification;
 use \Civi\ActionProvider\Utils\CustomField;
 
+use Civi\ActionProvider\Utils\Fields;
 use CRM_ActionProvider_ExtensionUtil as E;
 
 class GetEmailAddress extends AbstractAction {
-  
+
   /**
    * Run the action
-   * 
+   *
    * @param ParameterInterface $parameters
    *   The parameters to this action.
    * @param ParameterBagInterface $output
-   *   The parameters this action can send back 
+   *   The parameters this action can send back
    * @return void
    */
   protected function doAction(ParameterBagInterface $parameters, ParameterBagInterface $output) {
@@ -32,15 +33,15 @@ class GetEmailAddress extends AbstractAction {
       foreach($existingEmail as $field => $value) {
         $output->setParameter($field, $value);
       }
-      
+
     } catch (\Exception $e) {
       // Do nothing
     }
   }
-  
+
   /**
    * Returns the specification of the configuration options for the actual action.
-   * 
+   *
    * @return SpecificationBag
    */
   public function getConfigurationSpecification() {
@@ -51,10 +52,10 @@ class GetEmailAddress extends AbstractAction {
       new Specification('location_type_id', 'Integer', E::ts('Location type'), true, $defaultLocationType, null, $locationTypes, FALSE)
     ));
   }
-  
+
   /**
    * Returns the specification of the parameters of the actual action.
-   * 
+   *
    * @return SpecificationBag
    */
   public function getParameterSpecification() {
@@ -62,54 +63,18 @@ class GetEmailAddress extends AbstractAction {
       new Specification('contact_id', 'Integer', E::ts('Contact ID'), true),
     ));
   }
-  
+
   /**
    * Returns the specification of the output parameters of this action.
-   * 
+   *
    * This function could be overriden by child classes.
-   * 
+   *
    * @return SpecificationBag
    */
   public function getOutputSpecification() {
     $bag = new SpecificationBag();
-    $contact_fields = civicrm_api3('Email', 'getfields', array('action' => 'get', 'options' => array('limit' => 0)));
-    foreach($contact_fields['values'] as $field) {
-      if (empty($field['type'])) {
-        continue;
-      }
-      $type = \CRM_Utils_Type::typeToString($field['type']);
-      if (empty($type)) {
-        continue;
-      }
-      switch ($type) {
-        case 'Int':
-          $type = 'Integer';
-          break;
-      } 
-      
-      if (stripos($field['name'], 'custom_') === 0) {
-        // It is a custom field
-        $customFieldId = str_replace("custom_", "", $field['name']);
-        $fieldName = CustomField::getCustomFieldName($customFieldId);
-        $fieldSpec = new Specification(
-          $fieldName,
-          $type,
-          $field['title'],
-          false
-        );
-        $fieldSpec->setApiFieldName($field['name']);
-      } else {
-        $fieldSpec = new Specification(
-          $field['name'],
-          $type,
-          $field['title'],
-          false
-        );
-      }
-      $bag->addSpecification($fieldSpec);
-    }
-    
+    Fields::getFieldsForEntity($bag,'Email', 'get', array());
     return $bag;
   }
-  
+
 }