+++ Civi/ActionProvider/Utils/Type.php
@@ -1,0 +1,29 @@
+<?php
+/**
+ * @author Jaap Jansma <jaap.jansma@civicoop.org>
+ * @license AGPL-3.0
+ */
+
+namespace Civi\ActionProvider\Utils;
+
+class Type {
+
+  public static function convertCrmType($type) {
+    switch ($type) {
+      case 'Int':
+      case 'File':
+      case 'ContactReference':
+        return 'Integer';
+        break;
+      case 'Memo':
+        return 'Text';
+        break;
+      case 'StateProvince':
+      case 'Link':
+        return 'String';
+        break;
+    }
+    return $type;
+  }
+
+}